+++ CarRent.API/Domain/Entities/PaymentReceipt.cs
@@ -6,7 +6,7 @@ namespace CarRent.API.Domain.Entities
     {
         public int Id { get; set; }
         public virtual required Rental Rental { get; set; }
-        public double RentValue { get; set; }
+        public decimal RentValue { get; set; }
         public required string Observation { get; set; }
         public DateTime Emission { get; set; }
